Adding --debug and moving loading logic to load.go
