+++ cmd/pipeline/server/load.go
@@ -1,0 +1,60 @@
+package server
+
+import (
+	"log"
+	"strings"
+
+	"github.com/nats-io/nats"
+	"gopkg.in/yaml.v2"
+)
+
+type dataType struct {
+	Name  string            `yaml:"Name"`
+	Type  string            `yaml:"Type"`
+	CMD   string            `yaml:"CMD,omitempty"`
+	Nodes map[string]string `yaml:"Nodes,omitempty"`
+	In    map[string]string `yaml:"In,omitempty"`
+}
+
+// handleLoad downloads and processes a pipeline configuration file
+func (s *server) handleLoad(m *nats.Msg) {
+	// TODO: accept URL addresses
+	// TODO: Parse bitbucket requests... convert a to b
+	//   a: bitbucket.org/bign8/pipelines/sample/web
+	//   b: https://bitbucket.org/bign8/pipelines/raw/master/sample/web/pipeline.yml
+	// TODO: Parse github requests... convert a to b
+	//   a: github.com/bign8/pipelines/sample/web
+	//   b: https://github.com/bign8/pipelines/raw/master/sample/web/pipeline.yml
+	// log.Printf("Loading Config: %s", m.Data)
+	// resp, err := http.Get("http://" + string(m.Data) + "/pipeline.yml")
+	// if err != nil {
+	// 	log.Printf("Cannot Load: %s", m.Data)
+	// 	return
+	// }
+	// config, err := ioutil.ReadAll(resp.Body)
+
+	nodes := make(map[string]dataType)
+
+	// Parse YAML into memory structure
+	configData := strings.Split(string(m.Data), "---")[1:]
+	for _, configFile := range configData {
+		var config dataType
+		if err := yaml.Unmarshal([]byte("---\n"+configFile), &config); err != nil {
+			log.Printf("error loading config: %s", err)
+			return
+		}
+		if config.Type == "Node" {
+			nodes[config.Name] = config
+		}
+	}
+
+	// Initialize nodes into the server
+	for name, config := range nodes {
+		node := NewNode(name, config)
+		for streamName := range config.In {
+			s.Streams[streamName] = append(s.Streams[streamName], node)
+		}
+	}
+
+	log.Printf("Config: %+v\n", s.Streams)
+}
